Refactoring into smaller chunks of code
